refactor: extract agent dev tools from AgentWorkspace
